backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pyyaml==6.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0
apscheduler==3.10.4
python-telegram-bot==21.0
psycopg2-binary==2.9.9
